Time: 82 ms (94.95%), Space: 93.3 MB (27.55%) - LeetHub

+++ 2310-minimum-operations-to-halve-array-sum/2310-minimum-operations-to-halve-array-sum.cpp
@@ -1,22 +1,22 @@
 class Solution {
 public:
     int halveArray(vector<int>& nums) {
-         priority_queue<long double> pq;
-        long double sum = 0;
+         priority_queue< double> pq;
+         double sum = 0;
         
         for (const auto& it : nums) {
             sum += it;
             pq.push(it);
         }
 
-        long double target = sum / 2.0;
-        long double newSum = sum;
+        double target = sum / 2.0;
+         double newSum = sum;
         int operations = 0;
 
         while (newSum > target) {
-            long double element = pq.top();
+            double element = pq.top();
             pq.pop();
-            long double half = element / 2.0;
+            double half = element / 2.0;
             newSum -= half;
             pq.push(half);
             operations++;
